feat: add gui to warn user about the state of the asset in the build settings
Replace the content BuildIndexFromAsset to a static method with AssetBuildData struct to allow info gathering in property drawer.
Fix setting scene reference to null by clearing asset path when changing to asset to null.
Add help boxes to inform the user of the build index and state of the asst in the build list.